planning: move trajectory_stitcher to common
