+++ modules/planning/common/BUILD
@@ -476,4 +476,21 @@ cc_library(
     ],
 )
 
+cc_library(
+    name = "trajectory_stitcher",
+    srcs = [
+        "trajectory_stitcher.cc",
+    ],
+    hdrs = [
+        "trajectory_stitcher.h",
+    ],
+    deps = [
+        "//modules/planning/common/trajectory:publishable_trajectory",
+        "//cyber/common:log",
+        "//modules/common/proto:pnc_point_proto",
+        "//modules/planning/common:planning_gflags",
+        "//modules/planning/reference_line",
+    ],
+)
+
 cpplint()
