+++ modules/planning/common/trajectory_stitcher.h
